requirements.txt
llama-index-core==0.11.0.post1
llama-index-embeddings-fastembed==0.2.0
llama-index-llms-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-storage-docstore-mongodb==0.2.0
llama-index-storage-kvstore-mongodb==0.2.0
llama-index-vector-stores-qdrant==0.3.0
motor==3.5.1
onnx==1.16.2
onnxruntime==1.19.0
openai==1.42.0
pymongo==4.8.0
python-dotenv==1.0.1
qdrant-client==1.11.0